+++ common/src/database/sqlite.rs
@@ -40,7 +40,7 @@ use crate::bookmark::BookmarkData;
 use crate::database::Database;
 use crate::heartbeat::{HeartbeatData, HeartbeatsCache};
 use crate::subscription::{
-    SubscriptionData, SubscriptionMachine, SubscriptionMachineState, SubscriptionStatsCounters,
+    SubscriptionData, SubscriptionMachine, SubscriptionMachineState, SubscriptionStatsCounters, ContentFormat,
 };
 
 use super::schema::{Migration, MigrationBase, Version};
@@ -185,6 +185,7 @@ fn row_to_subscription(row: &Row) -> Result<SubscriptionData, rusqlite::Error> {
             ));
         }
     };
+    let content_format = ContentFormat::from_str(row.get::<&str, String>("content_format")?.as_ref()).map_err(|_| rusqlite::Error::InvalidColumnType(12, "content_format".to_owned(), Type::Text))?;
     Ok(SubscriptionData::from(
         row.get("uuid")?,
         row.get("version")?,
@@ -198,6 +199,7 @@ fn row_to_subscription(row: &Row) -> Result<SubscriptionData, rusqlite::Error> {
         row.get("max_envelope_size")?,
         row.get("enabled")?,
         row.get("read_existing_events")?,
+        content_format,
         outputs,
     ))
 }
@@ -567,10 +569,12 @@ impl Database for SQLiteDatabase {
                 conn.execute(
                     r#"INSERT INTO subscriptions (uuid, version, name, uri, query,
                     heartbeat_interval, connection_retry_count, connection_retry_interval,
-                    max_time, max_envelope_size, enabled, read_existing_events, outputs)
+                    max_time, max_envelope_size, enabled, read_existing_events, content_format,
+                    outputs)
                     VALUES (:uuid, :version, :name, :uri, :query,
                         :heartbeat_interval, :connection_retry_count, :connection_retry_interval,
-                        :max_time, :max_envelope_size, :enabled, :read_existing_events, :outputs)
+                        :max_time, :max_envelope_size, :enabled, :read_existing_events,
+                        :content_format, :outputs)
                     ON CONFLICT (uuid) DO UPDATE SET 
                         version = excluded.version,
                         name = excluded.name,
@@ -583,6 +587,7 @@ impl Database for SQLiteDatabase {
                         max_envelope_size = excluded.max_envelope_size,
                         enabled = excluded.enabled,
                         read_existing_events = excluded.read_existing_events,
+                        content_format = excluded.content_format,
                         outputs = excluded.outputs"#,
                     named_params! {
                         ":uuid": subscription.uuid(),
@@ -597,6 +602,7 @@ impl Database for SQLiteDatabase {
                         ":max_envelope_size": subscription.max_envelope_size(),
                         ":enabled": subscription.enabled(),
                         ":read_existing_events": subscription.read_existing_events(),
+                        ":content_format": subscription.content_format().to_string(),
                         ":outputs": serde_json::to_string(subscription.outputs())?,
                     },
                 )